Conflicts:
	HoccerXO/Crypto/CCRSA.m

+++ HoccerXO/Crypto/CCRSA.m
@@ -18,6 +18,8 @@
 
 #import "OpenSSLCrypto.h"
 
+#define RSA_DEBUG NO
+
 @implementation CCRSA
 
 static const uint32_t PADDING = kSecPaddingPKCS1;
@@ -146,10 +148,10 @@ static CCRSA *instance;
                            &plainBufferSize);
     
     if (status != noErr) {
-        NSLog(@"RSA:decryptWithKey: Error decrypting, OSStatus = %d", (NSInteger)status);
+        NSLog(@"CCRSA:decryptWithKey: Error decrypting, OSStatus = %d", (NSInteger)status);
         NSNotification *notification = [NSNotification notificationWithName:@"decryptionError" object:self];
         [[NSNotificationCenter defaultCenter] postNotification:notification];
-        NSLog(@"%@", [NSThread callStackSymbols]);
+        if (RSA_DEBUG) NSLog(@"%@", [NSThread callStackSymbols]);
     }
     
     //NSLog(@"decoded %d bytes, status %d", (NSInteger)plainBufferSize, (NSInteger)status);
@@ -369,8 +371,8 @@ static CCRSA *instance;
     NSData * publicTagSpec = [self publicTagForKeyIdString:publicKeyIdString];
     NSData * privateTagSpec = [self privateTagForKeyIdString:publicKeyIdString];
     
-    NSLog(@"cloneKeyPairKeys, cloning private key to %@", [[NSString alloc] initWithData:privateTagSpec encoding:NSUTF8StringEncoding]);
-    NSLog(@"cloneKeyPairKeys, cloning public key to %@", [[NSString alloc] initWithData:publicTagSpec encoding:NSUTF8StringEncoding]);
+    if (RSA_DEBUG) NSLog(@"cloneKeyPairKeys, cloning private key to %@", [[NSString alloc] initWithData:privateTagSpec encoding:NSUTF8StringEncoding]);
+    if (RSA_DEBUG) NSLog(@"cloneKeyPairKeys, cloning public key to %@", [[NSString alloc] initWithData:publicTagSpec encoding:NSUTF8StringEncoding]);
     
     if (![self addPrivateKeyBits:privateBits withTag:privateTagSpec]) {
         return NO;
@@ -383,38 +385,38 @@ static CCRSA *instance;
 }
 
 - (NSData *) publicTagForKeyIdString:(NSString *) keyIdString {
-    // NSLog(@"publicTagForKeyIdString keyIdString = %@",keyIdString);
+    if (RSA_DEBUG) NSLog(@"publicTagForKeyIdString keyIdString = %@",keyIdString);
     
     NSString * tagStr = [NSString stringWithUTF8String:[publicTag bytes]];
     tagStr = [tagStr stringByAppendingString:@"."];
     tagStr = [tagStr stringByAppendingString:keyIdString];
     
     NSData * tag = [tagStr dataUsingEncoding:NSUTF8StringEncoding];
-    NSLog(@"publicTagForKeyIdString, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
+    if (RSA_DEBUG) NSLog(@"publicTagForKeyIdString, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
     return tag;
 }
 
 - (NSData *) privateTagForKeyIdString:(NSString *) keyIdString {
-    //NSLog(@"privateTagForKeyIdString keyIdString = %@",keyIdString);
+    if (RSA_DEBUG) NSLog(@"privateTagForKeyIdString keyIdString = %@",keyIdString);
     
     NSString * tagStr = [NSString stringWithUTF8String:[privateTag bytes]];
     tagStr = [tagStr stringByAppendingString:@"."];
     tagStr = [tagStr stringByAppendingString:keyIdString];
     
     NSData * tag = [tagStr dataUsingEncoding:NSUTF8StringEncoding];
-    NSLog(@"privateTagForKeyIdString, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
+    if (RSA_DEBUG) NSLog(@"privateTagForKeyIdString, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
     return tag;
 }
 
 - (NSData *) publicTagForPeer:(NSString *) peerName {
-    //NSLog(@"publicTagForPeer peerName = %@",peerName);
+    if (RSA_DEBUG) NSLog(@"publicTagForPeer peerName = %@",peerName);
     
     NSString * tagStr = [NSString stringWithUTF8String:[publicPeerTag bytes]];
     tagStr = [tagStr stringByAppendingString:@"."];
     tagStr = [tagStr stringByAppendingString:peerName];
     
     NSData * tag = [tagStr dataUsingEncoding:NSUTF8StringEncoding];
-    //NSLog(@"publicTagForPeer, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
+    if (RSA_DEBUG) NSLog(@"publicTagForPeer, result %@", [[NSString alloc] initWithData:tag encoding:NSUTF8StringEncoding]);
     return tag;
 }
 
@@ -662,7 +664,7 @@ static CCRSA *instance;
                 }
                 //NSString * prefixString = [[NSString alloc] initWithData:prefix encoding:NSUTF8StringEncoding];
                 //NSLog(@"result %d = %@",i,resultObj[i]);
-                NSLog(@"tag %d = ‘%@‘",i,tag);
+                if (RSA_DEBUG) NSLog(@"tag %d = ‘%@‘",i,tag);
                 //NSLog(@"priv prefix %d = ‘%@‘",i,privatePrefixString);
                 //NSLog(@"pub  prefix %d = ‘%@‘",i,publicPrefixString);
                 NSString * postFix;
@@ -690,7 +692,7 @@ static CCRSA *instance;
                 
             }
         }
-        NSLog(@"found=%@",found);
+        if (RSA_DEBUG) NSLog(@"found=%@",found);
         return found;
     } else {
         NSLog(@"findKeyPairsWithPrefix: error OSStatus %d",(int)status);
@@ -719,7 +721,7 @@ static CCRSA *instance;
     if (status == 0 && resultObj != nil) {
         // do something with the result
         int results = [resultObj count];
-        NSLog(@"deleteAllRSAKeys: found %d keys",results);
+        if (RSA_DEBUG) NSLog(@"deleteAllRSAKeys: found %d keys",results);
         for (int i = 0; i < results;++i ) {
             //NSLog(@"result %d = %@",i,resultObj[i]);
             NSData * atag = resultObj[i][@"atag"];
@@ -729,7 +731,7 @@ static CCRSA *instance;
                 [dict setObject:atag forKey:(__bridge id)kSecAttrApplicationTag];
                 [dict setObject:(__bridge id) kSecAttrKeyTypeRSA forKey:(__bridge id)kSecAttrKeyType];
                 OSStatus myStatus = SecItemDelete((__bridge CFDictionaryRef)dict);
-                //NSLog(@"SecItemDelete returned %ld on dict %@", myStatus, dict);
+                if (RSA_DEBUG) NSLog(@"SecItemDelete returned %ld on dict %@", myStatus, dict);
                 if (myStatus == 0) {
                     ++deleted;
                 }
@@ -738,7 +740,7 @@ static CCRSA *instance;
                 NSLog(@"Error no atag for item %d = %@",i,resultObj[i]);
             }
         }
-        NSLog(@"deleted %d keys",deleted);
+        NSLog(@"CCRA: deleted %d keys",deleted);
     } else {
         NSLog(@"cleanAllRSAKeys: find error OSStatus %d",(int)status);
     }
@@ -803,7 +805,6 @@ static NSString *pemPrivateFooter = @"-----END RSA PRIVATE KEY-----";
     [encKey appendBytes:builder length:j + 2];
     [encKey appendData:publicKeyBits];
     
-    //NSString *returnString = [NSString stringWithFormat:@"%@\n%@\n%@", x509PublicHeader, [encKey asBase64EncodedString:1], x509PublicFooter];
     NSString *returnString = [NSString stringWithFormat:@"%@\n%@\n%@\n", x509PublicHeader, [encKey asBase64EncodedString:1], x509PublicFooter];
     //NSLog(@"PEM formatted key:\n%@",returnString);
     
@@ -932,9 +933,9 @@ size_t encodeLength(unsigned char * buf, size_t length) {
     if (([pemPrivateKeyString rangeOfString:pemPrivateHeader].location != NSNotFound) && ([pemPrivateKeyString rangeOfString:pemPrivateFooter].location != NSNotFound))
     {
         NSString *strippedKey = [self extractPayloadFromPEM: pemPrivateKeyString header: pemPrivateHeader footer: pemPrivateFooter];
+        if (RSA_DEBUG) NSLog(@"strippedkey:\n%@",strippedKey);
         return [NSData dataWithBase64EncodedString:strippedKey];
     }
-    NSLog(@"importPrivateKeyBits: not in PEM format");
     return nil;
 }
 
@@ -944,8 +945,9 @@ size_t encodeLength(unsigned char * buf, size_t length) {
     NSUInteger start = headerRange.location + headerRange.length;
     NSRange payloadRange = NSMakeRange(start, footerRange.location - start);
 
-    return [[[pemString substringWithRange: payloadRange]
+    return [[[[pemString substringWithRange: payloadRange]
              stringByReplacingOccurrencesOfString: @"\n" withString: @""]
+             stringByReplacingOccurrencesOfString: @"\r" withString: @""]
              stringByReplacingOccurrencesOfString: @" "  withString: @""];
 }
 
@@ -1021,7 +1023,7 @@ size_t encodeLength(unsigned char * buf, size_t length) {
         NSData * modulus = [CCRSA getPublicKeyMod:keyBits];
         return modulus.length * 8;
     } @catch (NSException* ex) {
-        NSLog(@"getPublicKeySize exception: %@", ex);
+        NSLog(@"#ERROR: getPublicKeySize exception: %@", ex);
     }
     return 0;
 }
